+++ src/components/Sidenav/GroupOptionsUI.vue
@@ -10,26 +10,6 @@
       v-close-overlay
     >
       <q-item
-        v-if="$q.platform.is.mobile"
-        :to="{name: 'groupMembers', params: {groupId: currentGroupId}}"
-      >
-        <q-icon
-          size="1em"
-          name="fas fa-users fa-fw on-left"
-        />
-        {{ $t('GROUP.MEMBERS') }}
-      </q-item>
-      <q-item
-        v-if="$q.platform.is.mobile"
-        :to="{name: 'groupSettings', params: {groupId: currentGroupId}}"
-      >
-        <q-icon
-          size="1em"
-          name="fas fa-cog fa-fw on-left"
-        />
-        {{ $t('GROUP.SETTINGS') }}
-      </q-item>
-      <q-item
         :to="{name: 'groupEdit', params: {groupId: currentGroupId}}"
       >
         <q-icon
@@ -51,7 +31,6 @@
       </q-item>
 
       <q-item
-        v-if="$q.platform.is.desktop"
         :to="{name: 'groupPreview', params: {groupPreviewId: currentGroupId}}"
       >
         <q-icon
@@ -62,7 +41,6 @@
       </q-item>
 
       <q-item
-        v-if="$q.platform.is.desktop"
         :to="{name: 'groupInvitations', params: {groupId: currentGroupId}}"
       >
         <q-icon