+++ src/components/Sidenav/SidenavBox.vue
@@ -1,21 +1,28 @@
 <template>
-  <q-card class="no-shadow grey-border">
+  <div
+    class="k-sidenav-box"
+    :class="{ expandable }"
+  >
     <q-toolbar
       class="toolbar"
-      @click.self="$emit('toggle')"
+      inverted
+      @click.native.stop.self="$emit('toggle')"
     >
       <slot name="icon" />
-      <q-toolbar-title>
+      <q-toolbar-title
+        @click.native.stop.self="$emit('toggle')"
+      >
         <slot name="name" />
       </q-toolbar-title>
       <slot name="tools" />
       <q-btn
+        v-if="expandable"
         flat
         round
         dense
         size="sm"
         class="card-arrow"
-        @click="$emit('toggle')"
+        @click.stop="$emit('toggle')"
       >
         <q-icon
           name="fas fa-angle-down arrow"
@@ -28,7 +35,7 @@
         <slot />
       </div>
     </transition>
-  </q-card>
+  </div>
 </template>
 
 <script>
@@ -36,6 +43,7 @@ import { QSlideTransition, QCard, QToolbar, QToolbarTitle, QBtn, QIcon } from 'q
 export default {
   components: { QSlideTransition, QCard, QToolbar, QToolbarTitle, QBtn, QIcon },
   props: {
+    expandable: { default: true, type: Boolean },
     expanded: { default: true, type: Boolean },
   },
 }
@@ -45,6 +53,11 @@ export default {
 @import '~variables'
 @import '~slidetoggle'
 
+.k-sidenav-box
+  margin-bottom 20px
+  &.expandable
+    .toolbar
+      cursor pointer
 .toolbar
   min-height 40px
   height 40px