+++ src/components/Sidenav/SidenavStoreUI.vue
@@ -7,6 +7,9 @@
     <template slot="icon">
       <q-icon name="fas fa-fw fa-shopping-cart" />
     </template>
+    <template slot="name">
+      {{ $t('GROUP.CURRENT_STORE') }}
+    </template>
     <div
       slot="tools"
       class="tools"